Extract WinToastTemplate impl to a seperate file
